pyproject.toml
[project]
name = "pythiabns"
version = "0.1.0"
description = "PythiaBNS: A pipeline for Binary Neutron Star post-merger analysis"
authors = [
    {name = "Stamatis Vretinaris", email = "svretina@gmail.com"}
]
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
keywords = ["gravitational-waves", "post-merger", "astrophysics", "bilby", "BNS", "neutron-stars"]
repository = "https://github.com/svretina/pythiabns"
requires-python = ">=3.13"
dependencies = [
    "astropy (>=7.1.0)",
    "dill (>=0.4.0)",
    "dynesty (>=2.1.5)",
    "h5py (>=3.13.0)",
    "matplotlib (>=3.10.3)",
    "numpy (>=2.2.6)",
    "pandas (>=2.2.3)",
    "pocomc (>=1.2.6)",
    "zeus-mcmc (>=2.5.4)",
    "tempest-sampler (>=0.1.5)",
    "numpyro (>=0.17.0)",
    "jax (>=0.5.0)",
    "jaxlib (>=0.5.0)",
    "nutpie (>=0.14.3)",
    "blackjax (>=1.2.5)",
    "cmdstanpy (>=1.2.5)",
    "scipy (>=1.15.3)",
    "tqdm (>=4.67.1)",
    "bilby (>=2.5.1)",
    "pycbc (>=2.6.0)",
    "multiprocess (>=0.70.17)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "ipython>=9.9.0",
    "corner>=2.2.3",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings-python>=1.8.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.9.0",
    "mypy>=1.0.0",
]

[tool.poetry]
packages = [{include = "pythiabns", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.13"
strict = false
ignore_missing_imports = true
